Rimosso `GameController` e aggiornato il flusso principale del gioco spostando la logica in `GameManager`, migliorando la coesione e riducendo la ridondanza.
